+++ module/data/_module.mjs
@@ -1,9 +1,7 @@
-// Export Actors
 export {default as BoilerplateActorBase} from "./actor-base.mjs";
 export {default as BoilerplateCharacter} from "./character.mjs";
 export {default as BoilerplateNPC} from "./npc.mjs";
 
-// Export Items
 export {default as BoilerplateItemBase} from "./item-base.mjs";
 export {default as BoilerplateItem} from "./item.mjs";
 export {default as BoilerplateFeature} from "./feature.mjs";