+++ module/data/actor-base.mjs
@@ -13,13 +13,12 @@ export default class BoilerplateActorBase extends foundry.abstract.TypeDataModel
       value: new fields.NumberField({ ...requiredInteger, initial: 5, min: 0 }),
       max: new fields.NumberField({ ...requiredInteger, initial: 5 })
     });
-    schema.biography = new fields.StringField({ required: true, blank: true }); // equivalent to passing ({initial: ""}) for StringFields
+    schema.biography = new fields.StringField({ required: true, blank: true });
 
     return schema;
   }
 
   prepareBaseData() {
     super.prepareBaseData();
-    // Initialize data structures before derived calculations
   }
 }