+++ module/data/npc.mjs
@@ -9,13 +9,8 @@ export default class BoilerplateNPC extends BoilerplateActorBase {
 
     schema.cr = new fields.NumberField({ ...requiredInteger, initial: 1, min: 0 });
     schema.xp = new fields.NumberField({ ...requiredInteger, initial: 0, min: 0 });
-    
-    return schema
-  }
 
-  prepareBaseData() {
-    super.prepareBaseData();
-    // Initialize data structures before derived calculations
+    return schema;
   }
 
   prepareDerivedData() {