+++ module/data/spell.mjs
@@ -6,7 +6,9 @@ export default class BoilerplateSpell extends BoilerplateItemBase {
     const fields = foundry.data.fields;
     const schema = super.defineSchema();
 
-    schema.spellLevel = new fields.NumberField({ required: true, nullable: false, integer: true, initial: 1, min: 1, max: 9 });
+    schema.spellLevel = new fields.NumberField({
+      required: true, nullable: false, integer: true, initial: 1, min: 1, max: 9
+    });
 
     return schema;
   }